refactor: update structure of pages folder
